+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
 ### Added
+
 - Added `InlineScriptRegistry` for handling inline scripts, instead of using `wp_add_inline_script`
 - Added `scripts_attributes` to ScriptVariables along with other classes dealing with Variables