+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Added `InlineScriptRegistry` for handling inline scripts, instead of using `wp_add_inline_script`
+- Added filter `airfleet/framework/inline-script-registry/content` to modify inline script content before output
+- Added filter `airfleet/framework/inline-script-registry/data-attributes` to modify script data attributes before output 
+- Added filter `airfleet/framework/inline-script-registry/deps` to modify script dependencies before output
+
 ## [1.14.0] - 2025-05-06
 
 ### Added