+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [1.15.0] - 2025-06-18
+
 ### Added
 
 - Added `InlineScriptRegistry` for handling inline scripts, instead of using `wp_add_inline_script`
@@ -215,7 +217,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added Options utilities
 - Added Features classes
 
-[unreleased]: https://github.com/airfleet/airfleet-wordpress-framework-php/compare/1.14.0...main
+[unreleased]: https://github.com/airfleet/airfleet-wordpress-framework-php/compare/1.15.0...main
 [0.1.0]: https://github.com/airfleet/airfleet-wordpress-framework-php/releases/tag/0.1.0
 [0.2.0]: https://github.com/airfleet/airfleet-wordpress-framework-php/releases/tag/0.2.0
 [0.2.1]: https://github.com/airfleet/airfleet-wordpress-framework-php/releases/tag/0.2.1
@@ -248,3 +250,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [1.13.0]: https://github.com/airfleet/airfleet-wordpress-framework-php/releases/tag/1.13.0
 
 [1.14.0]: https://github.com/airfleet/airfleet-wordpress-framework-php/releases/tag/1.14.0
+
+[1.15.0]: https://github.com/airfleet/airfleet-wordpress-framework-php/releases/tag/1.15.0